requirements.txt
# TrustLens AI – Python dependencies
# Install: pip install -r requirements.txt
# For AI providers install extras: pip install -r requirements.txt && pip install openai anthropic

# ── Core ─────────────────────────────────────────────────────────────────────
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.30.0,<1.0
pydantic>=2.7,<3.0
pydantic-settings>=2.2,<3.0

# ── Async / HTTP ─────────────────────────────────────────────────────────────
httpx>=0.27,<1.0
aiosqlite>=0.20,<1.0

# ── Browser automation ───────────────────────────────────────────────────────
playwright>=1.44,<2.0

# ── HTML / Content parsing ───────────────────────────────────────────────────
beautifulsoup4>=4.12,<5.0
lxml>=5.0,<6.0

# ── Security / Crypto ───────────────────────────────────────────────────────
cryptography>=42.0,<44.0

# ── Domain analysis ─────────────────────────────────────────────────────────
tldextract>=5.1,<6.0
python-Levenshtein>=0.25,<1.0

# ── DNS ──────────────────────────────────────────────────────────────────────
dnspython>=2.6,<3.0

# ── Logging ──────────────────────────────────────────────────────────────────
structlog>=24.1,<25.0

# ── Image / Screenshot Similarity ───────────────────────────────────────────
Pillow>=10.3,<12.0
imagehash>=4.3,<5.0

# ── Optional: AI Providers (uncomment as needed) ─────────────────────────────
# openai>=1.30,<2.0
# anthropic>=0.28,<1.0
